conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 25.7.0
asgiref=3.8.1=py313hca03da5_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2025.12.2=hca03da5_0
django=5.2=py313hca03da5_0
django-debug-toolbar=6.1.0=pypi_0
djangorestframework=3.16.1=pyhd8ed1ab_0
expat=2.7.1=h313beb8_0
libcxx=20.1.8=h8869778_0
libffi=3.4.4=hca03da5_1
libmpdec=4.0.0=h80987f9_0
libzlib=1.3.1=h5f15de7_0
ncurses=6.5=hee39554_0
openssl=3.0.18=h9b4081a_0
pip=25.2=pyhc872135_0
python=3.13.7=hc28b8d9_100_cp313
python_abi=3.13=1_cp313
readline=8.3=h0b18652_0
setuptools=78.1.1=py313hca03da5_0
sqlite=3.50.2=h79febb2_1
sqlparse=0.5.2=py313hca03da5_0
tk=8.6.15=hcd8a7d5_0
typing_extensions=4.15.0=py313hca03da5_0
tzdata=2025b=h04d1e81_0
waitress=3.0.2=py313hca03da5_0
wheel=0.45.1=py313hca03da5_0
xz=5.6.4=h80987f9_1
zlib=1.3.1=h5f15de7_0
